address additonal pr feedback
